google cloud function files/requirement.txt
functions-framework==3.*
functions-framework
google-cloud-storage
tensorflow
Pillow
numpy